+++ packages/e2e.generated.api-namespace.dep/src/generatedNoCheck/OntologyMetadata.ts
@@ -1,2 +1,4 @@
-export type $ExpectedClientVersion = '2.0.0';
+export type $ExpectedClientVersion = '2.1.0';
 export const $osdkMetadata = { extraUserAgent: 'typescript-sdk/dev osdk-cli/dev' };
+
+export const $ontologyRid = 'ri.ontology.main.ontology.dep';