+++ packages/e2e.generated.api-namespace.dep/src/generatedNoCheck/ontology/objects.ts
@@ -1,1 +1,1 @@
-export * from './objects/Task.js';
+export { Task } from './objects/Task.js';