+++ packages/e2e.generated.catchall/src/generatedNoCheck/ontology/queries.ts
@@ -1,2 +1,3 @@
-export * from './queries/getTodoCount.js';
-export * from './queries/queryTakesAllParameterTypes.js';
+export { getNamesOfCustomersFromCountry } from './queries/getNamesOfCustomersFromCountry.js';
+export { getTodoCount } from './queries/getTodoCount.js';
+export { queryTakesAllParameterTypes } from './queries/queryTakesAllParameterTypes.js';