+++ packages/e2e.sandbox.todoapp/src/generatedNoCheck2/index.ts
@@ -1,9 +1,9 @@
-export * from './ontology/actions';
+export { completeTodo, createTodo } from './ontology/actions';
 export * as $Actions from './ontology/actions';
-export * from './ontology/interfaces';
+export {} from './ontology/interfaces';
 export * as $Interfaces from './ontology/interfaces';
-export * from './ontology/objects';
+export { Todo } from './ontology/objects';
 export * as $Objects from './ontology/objects';
-export * from './ontology/queries';
+export {} from './ontology/queries';
 export * as $Queries from './ontology/queries';
 export { $ontologyRid } from './OntologyMetadata';