+++ packages/generator/src/GenerateContext/EnhanceCommon.ts
@@ -17,8 +17,6 @@
 import type { EnhancedOntologyDefinition } from "./EnhancedOntologyDefinition.js";
 
 export interface EnhanceCommon {
-  ontologyApiNamespace: string | undefined;
-  apiNamespacePackageMap: Map<string, string>;
   importExt: string;
   /** DO NOT USE THIS IN THE CONSTRUCTOR(S) */
   enhancedOntology: EnhancedOntologyDefinition;